Split base and skip traversal into separate classes
